auth backend/frontend init

+++ frontend/src/components/workout.jsx
@@ -1,25 +1,43 @@
+import { useAuthContext } from "../hooks/useAuthContext";
 import { useWorkoutsContext } from "../hooks/useWorkoutsContext";
-import formatDistanceToNow from 'date-fns/formatDistanceToNow';
+import formatDistanceToNow from "date-fns/formatDistanceToNow";
 
 export default function Workout({ workout }) {
-    const {dispatch} = useWorkoutsContext();
-    const delWorkout = async () => {
-        await fetch(`/api/workouts/${workout._id}`, {
-            method: 'DELETE'
-        }).then(res=>res.json()).then(res=>{
-            if(!res.error) {
-                dispatch({type: 'DELETE_WORKOUT', payload: res});
-            }
-            console.log(res);
-        }).catch(e=>console.log(e));
+  const { user } = useAuthContext();
+  const { dispatch } = useWorkoutsContext();
+  const delWorkout = async () => {
+    if (user) {
+      await fetch(`/api/workouts/${workout._id}`, {
+        method: "DELETE",
+        headers: {
+          Authorization: `Bearer ${user.token}`
+        }
+      })
+        .then((res) => res.json())
+        .then((res) => {
+          if (!res.error) {
+            dispatch({ type: "DELETE_WORKOUT", payload: res });
+          }
+          console.log(res);
+        })
+        .catch((e) => console.log(e));
     }
-    return (
-        <div className="workout-details">
-            <h4>{workout.title}</h4>
-            <p><strong>Load (kg): </strong> {workout.load}</p>
-            <p><strong>Reps (kg): </strong> {workout.reps}</p>
-            <p>{formatDistanceToNow(new Date(workout.createdAt), {addSuffix: true})}</p>
-            <span className="material-symbols-outlined" onClick={delWorkout}>Delete</span>
-        </div>
-    );
+  };
+  return (
+    <div className="workout-details">
+      <h4>{workout.title}</h4>
+      <p>
+        <strong>Load (kg): </strong> {workout.load}
+      </p>
+      <p>
+        <strong>Reps (kg): </strong> {workout.reps}
+      </p>
+      <p>
+        {formatDistanceToNow(new Date(workout.createdAt), { addSuffix: true })}
+      </p>
+      <span className="material-symbols-outlined" onClick={delWorkout}>
+        Delete
+      </span>
+    </div>
+  );
 }
